log and variable, function name changes; moved old deploment cleanup to helpers
